Added a kickstart for stalled solver
-If the solver fails, the game will now make a random move to kickstart
it into solving. This is a kludge- the solver routinely fails on the
last 6 even though it finds the solution after one move is made.
-Also fixed bug where you could crash the game by clicking solve it for
me before the preview countdown finished.
-Also started accounting for the movemaker catching up to the solver.
Now if the movemaker finishes before the solution is finished,
processSolvedGoal will restart the movemaker after it adds new moves to
the MoveQueue